Add detail to timeseriesByLocation.
