+++ src/scheduled/reports/_utils.js
@@ -1,0 +1,9 @@
+
+function uniqueByKey (arr, key) {
+  const h = arr.reduce((hsh, m) => { return { ...hsh, [m[key]]: m } }, {})
+  return Object.values(h)
+}
+
+module.exports = {
+  uniqueByKey
+}
